Widen the corpse scan's counter past a byte
The loop compares a `byte` against maxclients, and MAX_CLIENTS is 256 -- a
value the engine's coop clamp accepts. At exactly 256 the counter wraps
255 -> 0 and the condition never fails, hanging the game library inside the
server's frame. The indices stay in range, so it is a hang rather than a
memory error.

Server cvar only; a client cannot reach it.

+++ src/g_ai.c
@@ -1684,7 +1684,7 @@ void AI_AfterLife(edict_t *self)
 		edict_t	*other=NULL;
 		trace_t tr;
 		vec3_t	end, travelvec;
-		byte	i;
+		int		i;
 		long	len;
 		float	dir;
 		int     dir2;
